[VideoPlayer] [HLS] playlist tag の属性から値を取り出しやすくした
